Add context propagation
New features and breaking changes:

1. HTTP context propagation

The wrapper now automatically tries to extract B3 tracing headers from
incoming HTTP requests and uses the extracted span context as the parent
span when creating new spans.

2. Allows users to initialize the tracing without auto-creating the span

The decorator now must be called and accepts an optional argument
`with_span=<bool>` which defaults to `True`. When set to `False`, the
wrapper still initializes and flushes the tracer but does not auto
create spans.

3. Removed signalfx_lambda.wrapper

Since we are breaking backward compatibility here and releaing a
breaking version according to semver, we also removed the depracated
wrapper decorator.